backend/requirements.txt
Flask
Flask-Cors
Werkzeug
google-generativeai